[ui] refactor, fetch less
